Add files via upload
The project was done using Waikato Environment for Knowledge Analysis (WEKA).
The software contains a user interface that allows machine learning techniques to be utilized without implementing any coding. As a first project, this was an opportunity to learn the theory related to machine learning practices rather than the technical implementations through coding.